Extract product fetching from command line code.
Utilize the fetch package code for product fetching instead of having
the command line package handle it.
