+++ .travis.yml
@@ -18,4 +18,4 @@ deploy:
   skip_cleanup: true
   on:
       tags: true
-      # repo: Akenaide/wsoffcli
+      # repo: kwadkore/wsoffcli
