+++ fetch/products_test.go
@@ -1,4 +1,4 @@
-package cmd
+package fetch
 
 import (
 	"strings"
